requirements.txt
aiogram==2.10.1
aiohttp==3.7.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
gino==1.0.1
idna==2.10
importlib-metadata==1.7.0
jsonpickle==1.4.1
multidict==5.0.0
python-dotenv==0.14.0
pytz==2020.1
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
yarl==1.6.2
zipp==3.4.0
